v0.4

+++ cleanup.sh
@@ -1,7 +1,18 @@
 #!/bin/bash
 
-docker stop $(docker ps -qa)
-docker rm $(docker ps -qa)
+containers=$(docker ps -qa |wc -l)
+echo $containers
+
+if [ $containers -ge 1 ]; then
+  docker stop $(docker ps -qa)
+  docker rm $(docker ps -qa)
+fi
+
+none_images=$(docker images -a|awk '/none/ {print $3}')
+
+for image in $none_images; do
+  docker rmi $image
+done
 
 #docker rmi $(docker images -qa)
